Add Exceptions

+++ Meetup.Application/CommandsHandlers/DeleteEventCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using MediatR;
 using Meetup.Application.Commands;
+using Meetup.Application.Exceptions;
 using Meetup.Data.Interfaces;
+using Meetup.Models;
 
 namespace Meetup.Application.CommandsHandlers
 {
@@ -16,6 +18,11 @@ namespace Meetup.Application.CommandsHandlers
         public async Task<Unit> Handle(DeleteEventCommand request, CancellationToken cancellationToken)
         {
             var eventForDelete = await _eventRepository.GetByIdAsync(request.Id);
+
+            if (eventForDelete == null || eventForDelete.Id != request.Id)
+            {
+                throw new NotFoundException(nameof(Event), request.Id);
+            }
 
             _eventRepository.Delete(eventForDelete);
             await _eventRepository.SaveAsync();
